+++ src/core/Scene.h
@@ -48,7 +48,7 @@ public:
     // Serialization
     void saveScene(const std::string &name);
 
-    void loadScene(const std::string &name);
+    void loadScene(const std::string &name, const std::string &filepath = "");
 
 protected:
     std::string _name;
